LPS-184438 Move to proper package
