+++ modules/apps/headless/headless-user-notification/headless-user-notification-impl/src/main/java/com/liferay/headless/user/notification/internal/resource/v1_0/UserNotificationResourceImpl.java
@@ -182,7 +182,7 @@ public class UserNotificationResourceImpl
 	private ModelResourcePermission<User> _userModelResourcePermission;
 
 	@Reference(
-		target = "(component.name=com.liferay.headless.user.notification.internal.dto.v1_0.UserNotificationDTOConverter)"
+		target = "(component.name=com.liferay.headless.user.notification.internal.dto.v1_0.converter.UserNotificationDTOConverter)"
 	)
 	private DTOConverter<UserNotificationEvent, UserNotification>
 		_userNotificationDTOConverter;
